--- scheduler/pt_signin.py
@@ -1,26 +1,0 @@
-from threading import Lock
-
-import log
-from pt.signin import SignIn
-
-lock = Lock()
-
-
-class PTSignin:
-    signin = None
-
-    def __init__(self):
-        self.signin = SignIn()
-
-    def run_schedule(self):
-        """
-        运行PT站签到定时服务
-        """
-        try:
-            lock.acquire()
-            if self.signin:
-                self.signin.signin()
-        except Exception as err:
-            log.error("【RUN】执行任务pt_signin出错：%s" % str(err))
-        finally:
-            lock.release()
