+++ synology/INFO
@@ -1,5 +1,5 @@
 package="nastool"
-version="0.9.07"
+version="0.9.08"
 description="NAS资源归集整理工具：https://github.com/jxxghp/nas-tools"
 arch="noarch"
 distributor="jxxghp"
